Move Context Menu to separate component
